requirements.txt
playwright
Pillow
ollama                
pydantic    